requirements.txt
flask
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy==2.*
email_validator
python-decouple
gunicorn
SQLAlchemy==1.3.17
psycopg2==2.8.5
bip_utils==2.4.0
httpx==0.18.2
cryptography==37.0.4
pycryptodomex==3.15.0
python-dateutil~=2.8.2
xrpl-py==2.3.0